final-yessmine

+++ src/Controller/ModuleDocuments/BackOffice/DefaultGalleryController.php
@@ -3,8 +3,6 @@
 namespace App\Controller\ModuleDocuments\BackOffice;
 
 use App\Entity\DefaultGallery;
-use App\Entity\Family;
-use App\Entity\User;
 use App\Form\ModuleDocuments\BackOffice\DefaultGalleryType;
 use App\Repository\DefaultGalleryRepository;
 use Doctrine\ORM\EntityManagerInterface;
@@ -12,15 +10,14 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
-use App\Service\ActiveFamilyResolver;
 
 #[Route('/portal/admin/documents/default-galleries')]
 final class DefaultGalleryController extends AbstractController
 {
     #[Route(name: 'app_default_gallery_index', methods: ['GET'])]
-    public function index(DefaultGalleryRepository $defaultGalleryRepository, ActiveFamilyResolver $familyResolver): Response
+    public function index(DefaultGalleryRepository $defaultGalleryRepository): Response
     {
-        $this->resolveFamily($familyResolver);
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
 
         return $this->render('ModuleDocuments/BackOffice/default_gallery/index.html.twig', [
             'default_galleries' => $defaultGalleryRepository->findAll(),
@@ -28,9 +25,9 @@ final class DefaultGalleryController extends AbstractController
     }
 
     #[Route('/new', name: 'app_default_gallery_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager, ActiveFamilyResolver $familyResolver): Response
+    public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
-        $this->resolveFamily($familyResolver);
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
 
         $defaultGallery = new DefaultGallery();
         $form = $this->createForm(DefaultGalleryType::class, $defaultGallery);
@@ -50,9 +47,9 @@ final class DefaultGalleryController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_default_gallery_show', methods: ['GET'])]
-    public function show(DefaultGallery $defaultGallery, ActiveFamilyResolver $familyResolver): Response
+    public function show(DefaultGallery $defaultGallery): Response
     {
-        $this->resolveFamily($familyResolver);
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
 
         return $this->render('ModuleDocuments/BackOffice/default_gallery/show.html.twig', [
             'default_gallery' => $defaultGallery,
@@ -60,9 +57,9 @@ final class DefaultGalleryController extends AbstractController
     }
 
     #[Route('/{id}/edit', name: 'app_default_gallery_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, DefaultGallery $defaultGallery, EntityManagerInterface $entityManager, ActiveFamilyResolver $familyResolver): Response
+    public function edit(Request $request, DefaultGallery $defaultGallery, EntityManagerInterface $entityManager): Response
     {
-        $this->resolveFamily($familyResolver);
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
 
         $form = $this->createForm(DefaultGalleryType::class, $defaultGallery);
         $form->handleRequest($request);
@@ -80,9 +77,9 @@ final class DefaultGalleryController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_default_gallery_delete', methods: ['POST'])]
-    public function delete(Request $request, DefaultGallery $defaultGallery, EntityManagerInterface $entityManager, ActiveFamilyResolver $familyResolver): Response
+    public function delete(Request $request, DefaultGallery $defaultGallery, EntityManagerInterface $entityManager): Response
     {
-        $this->resolveFamily($familyResolver);
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
 
         if ($this->isCsrfTokenValid('delete'.$defaultGallery->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($defaultGallery);
@@ -90,20 +87,5 @@ final class DefaultGalleryController extends AbstractController
         }
 
         return $this->redirectToRoute('app_default_gallery_index', [], Response::HTTP_SEE_OTHER);
-    }
-
-    private function resolveFamily(ActiveFamilyResolver $familyResolver): Family
-    {
-        $user = $this->getUser();
-        if (!$user instanceof User) {
-            throw $this->createAccessDeniedException();
-        }
-
-        $family = $familyResolver->resolveForUser($user);
-        if ($family === null) {
-            throw $this->createAccessDeniedException();
-        }
-
-        return $family;
     }
 }
